+++ packages/react-await/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - 837ed18: feat(\*): remove target react v16, v17
 
+## 0.0.16
+
+### Patch Changes
+
+- a6d279e: fix(\*): add use-sync-external-store
+
 ## 0.0.15
 
 ### Patch Changes